+++ src/api/java/blusunrize/immersiveengineering/api/crafting/RefineryRecipe.java
@@ -18,6 +18,7 @@ import net.minecraftforge.fluids.FluidStack;
 import net.minecraftforge.registries.RegistryObject;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import java.util.Optional;
 
 /**
@@ -32,10 +33,11 @@ public class RefineryRecipe extends MultiblockRecipe
 
 	public final FluidStack output;
 	public final FluidTagInput input0;
+	@Nullable
 	public final FluidTagInput input1;
 	public final Ingredient catalyst;
 
-	public RefineryRecipe(ResourceLocation id, FluidStack output, FluidTagInput input0, FluidTagInput input1, Ingredient catalyst, int energy)
+	public RefineryRecipe(ResourceLocation id, FluidStack output, FluidTagInput input0, @Nullable FluidTagInput input1, Ingredient catalyst, int energy)
 	{
 		super(LAZY_EMPTY, IERecipeTypes.REFINERY, id);
 		this.output = output;
@@ -44,7 +46,9 @@ public class RefineryRecipe extends MultiblockRecipe
 		this.catalyst = catalyst;
 		setTimeAndEnergy(1, energy);
 
-		this.fluidInputList = Lists.newArrayList(this.input0, this.input1);
+		this.fluidInputList = Lists.newArrayList(this.input0);
+		if(this.input1!=null)
+			this.fluidInputList.add(this.input1);
 		this.fluidOutputList = Lists.newArrayList(this.output);
 	}
 
@@ -54,27 +58,27 @@ public class RefineryRecipe extends MultiblockRecipe
 		return SERIALIZER.get();
 	}
 
-	public static RefineryRecipe findRecipe(Level level, FluidStack input0, FluidStack input1, ItemStack catalyst)
+	public static RefineryRecipe findRecipe(Level level, FluidStack input0, @Nonnull FluidStack input1, @Nonnull ItemStack catalyst)
 	{
 		for(RefineryRecipe recipe : RECIPES.getRecipes(level))
 		{
 			if(!recipe.catalyst.test(catalyst))
 				continue;
-			if(input0!=null)
+			if(!input0.isEmpty())
 			{
 				if(recipe.input0!=null&&recipe.input0.test(input0))
 				{
-					if((recipe.input1==null&&input1==null)||(recipe.input1!=null&&input1!=null&&recipe.input1.test(input1)))
+					if((recipe.input1==null&&input1.isEmpty())||(recipe.input1!=null&&recipe.input1.test(input1)))
 						return recipe;
 				}
 
 				if(recipe.input1!=null&&recipe.input1.test(input0))
 				{
-					if((recipe.input0==null&&input1==null)||(recipe.input0!=null&&input1!=null&&recipe.input0.test(input1)))
+					if((recipe.input0==null&&input1.isEmpty())||(recipe.input0!=null&&recipe.input0.test(input1)))
 						return recipe;
 				}
 			}
-			else if(input1!=null)
+			else if(!input1.isEmpty())
 			{
 				if(recipe.input0!=null&&recipe.input0.test(input1)&&recipe.input1==null)
 					return recipe;
@@ -93,16 +97,16 @@ public class RefineryRecipe extends MultiblockRecipe
 		{
 			if(!input0.isEmpty()&&input1.isEmpty())
 			{
-				if(recipe.input0.testIgnoringAmount(input0)||recipe.input1.testIgnoringAmount(input0))
+				if(recipe.input0.testIgnoringAmount(input0)||(recipe.input1!=null&&recipe.input1.testIgnoringAmount(input0)))
 					return Optional.of(recipe);
 			}
 			else if(input0.isEmpty()&&!input1.isEmpty())
 			{
-				if(recipe.input0.testIgnoringAmount(input1)||recipe.input1.testIgnoringAmount(input1))
+				if(recipe.input0.testIgnoringAmount(input1)||(recipe.input1!=null&&recipe.input1.testIgnoringAmount(input1)))
 					return Optional.of(recipe);
 			}
-			else if((recipe.input0.testIgnoringAmount(input0)&&recipe.input1.testIgnoringAmount(input1))
-					||(recipe.input1.testIgnoringAmount(input0)&&recipe.input0.testIgnoringAmount(input1)))
+			else if((recipe.input1!=null&&recipe.input0.testIgnoringAmount(input0)&&recipe.input1.testIgnoringAmount(input1))
+					||(recipe.input1!=null&&recipe.input1.testIgnoringAmount(input0)&&recipe.input0.testIgnoringAmount(input1)))
 				return Optional.of(recipe);
 		}
 		return Optional.empty();