+++ src/datagen/java/blusunrize/immersiveengineering/data/tags/IEBlockTags.java
@@ -31,6 +31,7 @@ import net.minecraft.data.DataGenerator;
 import net.minecraft.data.tags.BlockTagsProvider;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.BlockTags;
+import net.minecraft.tags.ItemTags;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.item.Tiers;
@@ -66,6 +67,7 @@ public class IEBlockTags extends BlockTagsProvider
 				.add(WoodenDecoration.TREATED_FENCE.get());
 		tag(BlockTags.WOODEN_FENCES)
 				.add(WoodenDecoration.TREATED_FENCE.get());
+		tag(BlockTags.PLANKS).add(WoodenDecoration.FIBERBOARD.get());
 		tag(IETags.fencesSteel)
 				.add(MetalDecoration.STEEL_FENCE.get());
 		tag(IETags.fencesAlu)