+++ src/datagen/java/blusunrize/immersiveengineering/data/tags/IEItemTags.java
@@ -15,10 +15,12 @@ import blusunrize.immersiveengineering.api.Lib;
 import blusunrize.immersiveengineering.api.utils.TagUtils;
 import blusunrize.immersiveengineering.common.blocks.metal.BasicConnectorBlock;
 import blusunrize.immersiveengineering.common.items.WireCoilItem;
+import blusunrize.immersiveengineering.common.register.IEBlocks;
 import blusunrize.immersiveengineering.common.register.IEBlocks.BlockEntry;
 import blusunrize.immersiveengineering.common.register.IEBlocks.Cloth;
 import blusunrize.immersiveengineering.common.register.IEBlocks.Connectors;
 import blusunrize.immersiveengineering.common.register.IEBlocks.MetalDevices;
+import blusunrize.immersiveengineering.common.register.IEBlocks.WoodenDecoration;
 import blusunrize.immersiveengineering.common.register.IEItems.*;
 import com.google.common.base.Preconditions;
 import net.minecraft.data.DataGenerator;
@@ -27,6 +29,7 @@ import net.minecraft.data.tags.ItemTagsProvider;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.ItemTags;
 import net.minecraft.world.item.Items;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.common.data.ExistingFileHelper;
 
@@ -102,6 +105,7 @@ public class IEItemTags extends ItemTagsProvider
 				.addTag(IETags.aluminumRod)
 				.addTag(IETags.ironRod)
 				.addTag(IETags.steelRod);
+		tag(IETags.plasticPlate).add(Ingredients.DUROPLAST_PLATE.asItem());
 		tag(IETags.sawblades).add(Tools.SAWBLADE.get());
 		tag(IETags.circuitPCB).add(Ingredients.CIRCUIT_BOARD.asItem());
 		tag(IETags.circuitLogic).add(Ingredients.ELECTRON_TUBE.asItem());
@@ -109,6 +113,9 @@ public class IEItemTags extends ItemTagsProvider
 		tag(IETags.hammers).add(Tools.HAMMER.get());
 		tag(IETags.screwdrivers).add(Tools.SCREWDRIVER.get());
 		tag(IETags.wirecutters).add(Tools.WIRECUTTER.get());
+		tag(IETags.connectorInsulator)
+				.add(Blocks.TERRACOTTA.asItem())
+				.add(IEBlocks.StoneDecoration.DUROPLAST.asItem());
 		tag(ItemTags.CLUSTER_MAX_HARVESTABLES).add(Tools.STEEL_PICK.get());
 
 		generateTagsForToolbox();