+++ src/main/java/blusunrize/immersiveengineering/common/register/IEBlocks.java
@@ -138,6 +138,10 @@ public final class IEBlocks
 				"coresample", STONE_DECO_PROPS_NOT_SOLID, p -> new HorizontalFacingBlock<>(IEBlockEntities.CORE_SAMPLE, p)
 		);
 
+		public static final BlockEntry<IEBaseBlock> DUROPLAST = BlockEntry.simple(
+				"duroplast", STONE_DECO_PROPS_NOT_SOLID
+		);
+
 		private static void init()
 		{
 		}
@@ -309,6 +313,10 @@ public final class IEBlocks
 						.strength(0.5F)
 						.noCollission().noOcclusion(),
 				SawdustBlock::new
+		);
+		public static final BlockEntry<IEBaseBlock> FIBERBOARD = BlockEntry.simple("fiberboard",
+				() -> Block.Properties.of(Material.WOOD, MaterialColor.WOOD)
+						.strength(1.25f, 1)
 		);
 
 		private static void init()
@@ -670,7 +678,8 @@ public final class IEBlocks
 
 		public static BlockEntry<IEBaseBlock> simple(String name, Supplier<Properties> properties)
 		{
-			return simple(name, properties, $ -> {});
+			return simple(name, properties, $ -> {
+			});
 		}
 
 		public static BlockEntry<IEEntityBlock<?>> barrel(String name, boolean metal)
@@ -678,7 +687,8 @@ public final class IEBlocks
 			return new BlockEntry<>(name, () -> BarrelBlock.getProperties(metal), p -> BarrelBlock.make(p, metal));
 		}
 
-		public static BlockEntry<ScaffoldingBlock> scaffolding(String name, Supplier<Properties> props) {
+		public static BlockEntry<ScaffoldingBlock> scaffolding(String name, Supplier<Properties> props)
+		{
 			return new BlockEntry<>(name, props, ScaffoldingBlock::new);
 		}
 
@@ -687,11 +697,13 @@ public final class IEBlocks
 			return new BlockEntry<>(name, props, FenceBlock::new);
 		}
 
-		public static BlockEntry<PostBlock> post(String name, Supplier<Properties> props) {
+		public static BlockEntry<PostBlock> post(String name, Supplier<Properties> props)
+		{
 			return new BlockEntry<>(name, props, PostBlock::new);
 		}
 
-		public static BlockEntry<WallmountBlock> wallmount(String name, Supplier<Properties> props) {
+		public static BlockEntry<WallmountBlock> wallmount(String name, Supplier<Properties> props)
+		{
 			return new BlockEntry<>(name, props, WallmountBlock::new);
 		}
 
@@ -712,7 +724,7 @@ public final class IEBlocks
 		public BlockEntry(BlockEntry<? extends T> toCopy)
 		{
 			this.properties = toCopy.properties;
-			this.regObject = (RegistryObject<T>) toCopy.regObject;
+			this.regObject = (RegistryObject<T>)toCopy.regObject;
 		}
 
 		@Override
@@ -721,11 +733,13 @@ public final class IEBlocks
 			return regObject.get();
 		}
 
-		public BlockState defaultBlockState() {
+		public BlockState defaultBlockState()
+		{
 			return get().defaultBlockState();
 		}
 
-		public ResourceLocation getId() {
+		public ResourceLocation getId()
+		{
 			return regObject.getId();
 		}