+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/jei/JEIHelper.java
@@ -62,6 +62,7 @@ import net.minecraftforge.registries.ForgeRegistries;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.List;
 import java.util.Optional;
 import java.util.function.Predicate;
 
@@ -239,7 +240,7 @@ public class JEIHelper implements IModPlugin
 				if(!bucket.isEmpty()&&tag.isPresent())
 					recipes.add(new BottlingMachineRecipe(
 							new ResourceLocation(Lib.MODID, "jei_bucket_"+Registry.FLUID.getKey(f).getPath()),
-							Lazy.of(() -> bucket),
+							List.of(Lazy.of(() -> bucket)),
 							Ingredient.of(Items.BUCKET),
 							new FluidTagInput(tag.get(), 1000)
 					));