+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/jei/bottlingmachine/BottlingMachineRecipeCategory.java
@@ -29,6 +29,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.fluids.FluidType;
 
 import java.util.Arrays;
+import java.util.List;
 
 public class BottlingMachineRecipeCategory extends IERecipeCategory<BottlingMachineRecipe>
 {
@@ -50,9 +51,12 @@ public class BottlingMachineRecipeCategory extends IERecipeCategory<BottlingMach
 		builder.addSlot(RecipeIngredientRole.INPUT, 1, 13)
 				.addItemStacks(Arrays.asList(recipe.input.getItems()))
 				.setBackground(JEIHelper.slotDrawable, -1, -1);
-		builder.addSlot(RecipeIngredientRole.OUTPUT, 101, 13)
-				.addItemStack(recipe.output.get())
-				.setBackground(JEIHelper.slotDrawable, -1, -1);
+
+		List<ItemStack> outputs = recipe.output.get();
+		for(int i=0; i<outputs.size(); i++)
+			builder.addSlot(RecipeIngredientRole.OUTPUT, 101, 13+i*18)
+					.addItemStack(outputs.get(i))
+					.setBackground(JEIHelper.slotDrawable, -1, -1);
 
 		builder.addSlot(RecipeIngredientRole.INPUT, 76, 1)
 				.setFluidRenderer(4*FluidType.BUCKET_VOLUME, false, 16, 48)